Improve matching limit implementation

+++ java/src/cz/iocb/sachem/molecule/NativeIsomorphism.java
@@ -28,13 +28,13 @@ public class NativeIsomorphism
     }
 
 
-    public float match(byte[] query, int limit) throws IterationLimitExceededException
+    public float match(byte[] query, long limit) throws IterationLimitExceededException
     {
         return match(implementation, query, limit);
     }
 
 
-    private static native float match(ByteBuffer implementation, byte[] query, int limit)
+    private static native float match(ByteBuffer implementation, byte[] query, long limit)
             throws IterationLimitExceededException;
